+++ src/rpc.cpp
@@ -698,10 +698,9 @@ json RPCServer::serialize_json_value(JSValueRef js_value) {
     }
 #if REALM_ENABLE_SYNC
     else if (jsc::Object::is_instance<js::UserClass<jsc::Types>>(m_context, js_object)) {
-        auto user = *jsc::Object::get_internal<js::UserClass<jsc::Types>>(js_object);
+        std::shared_ptr<SyncUser> user = *jsc::Object::get_internal<js::UserClass<jsc::Types>>(js_object);
         json user_dict {
             {"identity", user->identity()},
-            {"server", user->server_url()},
         };
         return {
             {"type", RealmObjectTypesUser},